+++ travel-with-me/src/main/java/com/frog/travelwithme/global/config/PasswordEncoderConfig.java
@@ -1,4 +1,4 @@
-package com.frog.travelwithme.common.config;
+package com.frog.travelwithme.global.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
